+++ src/main/java/com/example/Dummy2.java
@@ -1,9 +1,8 @@
-package com.example.demo;
+package com.example;
 
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.Step;
 import org.springframework.batch.core.configuration.annotation.JobBuilderFactory;
-import org.springframework.batch.core.configuration.annotation.JobScope;
 import org.springframework.batch.core.configuration.annotation.StepBuilderFactory;
 import org.springframework.batch.repeat.RepeatStatus;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,7 +10,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class Dummy1 {
+public class Dummy2 {
 
     @Autowired
     JobBuilderFactory jobs;
@@ -20,16 +19,16 @@ public class Dummy1 {
 
     @Bean
     public Step step1() {
-        return steps.get("dummy-tasklet-1")
+        return steps.get("dummy-tasklet-2")
                 .tasklet(((stepContribution, chunkContext) -> {
-                    System.out.println("hello 1");
+                    System.out.println("hello 2");
                     return RepeatStatus.FINISHED;
                 })).build();
     }
 
     @Bean
     public Job job() {
-        return jobs.get("dummy-job-1")
+        return jobs.get("dummy-job-2")
                 .start(step1())
                 .build();
     }
